--- a/src/python/pants/option/global_options.py
+++ b/src/python/pants/option/global_options.py
@@ -1510,25 +1510,6 @@ class GlobalOptions(BootstrapOptions, Subsystem):
         metavar="[+-]tag1,tag2,...",
     )
 
-    files_not_found_behavior = EnumOption(
-        "--files-not-found-behavior",
-        default=UnmatchedBuildFileGlobs.warn,
-        help=softwrap(
-            """
-            What to do when files and globs specified in BUILD files, such as in the
-            `sources` field, cannot be found. This happens when the files do not exist on
-            your machine or when they are ignored by the `--pants-ignore` option.
-            """
-        ),
-        advanced=True,
-        removal_version="2.14.0.dev0",
-        removal_hint=softwrap(
-            """
-            Use `[GLOBAL].unmatched_build_file_globs` instead, which behaves the same. This
-            option was renamed for clarity with the new `[GLOBAL].unmatched_cli_globs` option.
-            """
-        ),
-    )
     unmatched_build_file_globs = EnumOption(
         "--unmatched-build-file-globs",
         default=UnmatchedBuildFileGlobs.warn,