In the process, this closes https://github.com/pantsbuild/pants/issues/15802. 

Before:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd --no-buildsense-enable --no-anonymous-telemetry-enabled --spec-files=tgts.txt list'
Benchmark 1: ./pants --no-pantsd --no-buildsense-enable --no-anonymous-telemetry-enabled --spec-files=tgts.txt list
  Time (mean ± σ):      4.255 s ±  0.034 s    [User: 3.906 s, System: 0.545 s]
  Range (min … max):    4.223 s …  4.314 s    5 runs
```

After:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd --no-buildsense-enable --no-anonymous-telemetry-enabled --spec-files=tgts.txt list'
Benchmark 1: ./pants --no-pantsd --no-buildsense-enable --no-anonymous-telemetry-enabled --spec-files=tgts.txt list
  Time (mean ± σ):      3.837 s ±  0.271 s    [User: 3.315 s, System: 0.452 s]
  Range (min … max):    3.645 s …  4.302 s    5 runs
```